requirements.txt
discord.py==1.6.0
sentry-sdk==1.0.0
Pillow==8.1.2
